Fixes tests with IterableDataset
